Move blockchain to root level
Blockchain is an use case according to our architecture. Therefore, it
doesn't belong in the core package.
